+++ src/mm_animation_library.h
@@ -53,8 +53,8 @@ public:
     virtual ~MMAnimationLibrary();
     void bake_data(const AnimationMixer* p_player, const Skeleton3D* p_skeleton);
     MMQueryOutput query(const MMQueryInput& p_query_input);
-    size_t get_dim_count() const;
-    int32_t get_animation_pose_count(String p_animation_name) const;
+    int64_t get_dim_count() const;
+    int64_t get_animation_pose_count(String p_animation_name) const;
 
 #ifdef TOOLS_ENABLED
     void display_data(const Ref<EditorNode3DGizmo>& p_gizmo, const Transform3D& p_transform, String p_animation_name, int32_t p_pose_index) const;
